feat(mastracode): split rules from audit log
