+++ mastracode/web/src/web/ui/domains/factory/__tests__/FactoryPages.msw.test.tsx
@@ -500,7 +500,7 @@ function dragTo(card: HTMLElement, target: HTMLElement) {
 }
 
 describe('Factory sidebar section', () => {
-  it('given a GitHub project, when the app renders, then Factory exposes sibling Work and Review links', async () => {
+  it('given a GitHub project, when the app renders, then Factory exposes each workflow as a sibling link', async () => {
     useBoardHandlers();
     renderAt('/factory/work');
 
@@ -509,7 +509,8 @@ describe('Factory sidebar section', () => {
     expect(await within(nav).findByRole('link', { name: 'Work' })).toHaveAttribute('href', '/factory/work');
     expect(within(nav).getByRole('link', { name: 'Review' })).toHaveAttribute('href', '/factory/review');
     expect(within(nav).getByRole('link', { name: /Metrics/ })).toHaveAttribute('href', '/factory/metrics');
-    expect(within(nav).getByRole('link', { name: /Audit/ })).toHaveAttribute('href', '/factory/audit');
+    expect(within(nav).getByRole('link', { name: 'Rules' })).toHaveAttribute('href', '/factory/rules');
+    expect(within(nav).getByRole('link', { name: 'Audit log' })).toHaveAttribute('href', '/factory/audit');
     expect(within(nav).getByRole('region', { name: 'Factory sessions' })).toBeInTheDocument();
   });
 
@@ -528,7 +529,8 @@ describe('Factory sidebar section', () => {
     expect(within(nav).getByRole('link', { name: 'Work' })).toHaveAttribute('href', '/factory/work');
     expect(within(nav).getByRole('link', { name: 'Review' })).toHaveAttribute('href', '/factory/review');
     expect(within(nav).getByRole('link', { name: /Metrics/ })).toHaveAttribute('href', '/factory/metrics');
-    expect(within(nav).getByRole('link', { name: /Audit/ })).toHaveAttribute('href', '/factory/audit');
+    expect(within(nav).getByRole('link', { name: 'Rules' })).toHaveAttribute('href', '/factory/rules');
+    expect(within(nav).getByRole('link', { name: 'Audit log' })).toHaveAttribute('href', '/factory/audit');
   });
 });
 
@@ -558,7 +560,7 @@ describe('Factory workflow routing', () => {
     renderAt('/factory/work', localProject);
 
     expect(
-      await screen.findByText(/Board, metrics, and audit are available for server-backed Factories/),
+      await screen.findByText(/Board, metrics, rules, and audit are available for server-backed Factories/),
     ).toBeInTheDocument();
     expect(screen.queryByTestId('board-column-intake')).not.toBeInTheDocument();
   });
